Separated feature utils infra
