+++ ludwig/schema/features/utils.py
@@ -1,6 +1,10 @@
 from ludwig.schema import utils as schema_utils
 from ludwig.utils.registry import Registry
 
+from ludwig.decoders.registry import get_decoder_classes
+from ludwig.encoders.registry import get_encoder_classes
+from ludwig.schema.utils import create_cond
+
 input_type_registry = Registry()
 output_type_registry = Registry()
 
@@ -9,6 +13,7 @@ def register_input_feature(name: str):
     def wrap(cls):
         input_type_registry[name] = cls
         return cls
+
     return wrap
 
 
@@ -16,11 +21,19 @@ def register_output_feature(name: str):
     def wrap(cls):
         output_type_registry[name] = cls
         return cls
+
     return wrap
 
 
 def get_input_feature_jsonschema():
+    """
+    This function returns a JSON schema structured to only requires a `type` key and then conditionally applies a
+    corresponding input feature's field constraints.
 
+    Returns: JSON Schema
+
+    """
+    input_feature_types = sorted(list(input_type_registry.keys()))
     return {
         "type": "array",
         "items": {
@@ -31,15 +44,20 @@ def get_input_feature_jsonschema():
                 "column": {"type": "string"},
                 "encoder": {"type": "string"},
             },
-            "allOf": get_input_encoder_conds(input_feature_types)
-                     + get_input_preproc_conds(input_feature_types),
+            "allOf": get_encoder_conds(input_feature_types) + get_input_preproc_conds(input_feature_types),
             "required": ["name", "type"],
         }
     }
 
 
 def get_output_feature_jsonschema():
+    """
+    This function returns a JSON schema structured to only requires a `type` key and then conditionally applies a
+    corresponding output feature's field constraints.
 
+    Returns: JSON Schema
+    """
+    output_feature_types = sorted(list(output_type_registry.keys()))
     return {
         "type": "array",
         "items": {
@@ -50,14 +68,13 @@ def get_output_feature_jsonschema():
                 "column": {"type": "string"},
                 "decoder": {"type": "string"},
             },
-            "allOf": get_output_decoder_conds(output_feature_types)
-            + get_output_preproc_conds(output_feature_types),
+            "allOf": get_decoder_conds(output_feature_types) + get_output_preproc_conds(output_feature_types),
             "required": ["name", "type"],
         },
     }
 
 
-def get_input_encoder_conds(input_feature_types):
+def get_encoder_conds(input_feature_types):
     conds = []
     for feature_type in input_feature_types:
         encoder_names = list(get_encoder_classes(feature_type).keys())
@@ -69,10 +86,39 @@ def get_input_encoder_conds(input_feature_types):
     return conds
 
 
+def get_decoder_conds(input_feature_types):
+    conds = []
+    for feature_type in input_feature_types:
+        decoder_names = list(get_decoder_classes(feature_type).keys())
+        decoder_cond = create_cond(
+            {"type": feature_type},
+            {"decoder": {"enum": decoder_names}},
+        )
+        conds.append(decoder_cond)
+    return conds
+
+
 def get_input_preproc_conds(input_feature_types):
     conds = []
     for feature_type in input_feature_types:
         feature_cls = input_type_registry[feature_type]
+        preproc_spec = {
+            "type": "object",
+            "properties": feature_cls.preprocessing_schema(),
+            "additionalProperties": False,
+        }
+        preproc_cond = create_cond(
+            {"type": feature_type},
+            {"preprocessing": preproc_spec},
+        )
+        conds.append(preproc_cond)
+    return conds
+
+
+def get_output_preproc_conds(output_feature_types):
+    conds = []
+    for feature_type in output_feature_types:
+        feature_cls = output_type_registry[feature_type]
         preproc_spec = {
             "type": "object",
             "properties": feature_cls.preprocessing_schema(),
